+++ src/main/java/com/oskarro/justcode/repositories/ArticleRepository.java
@@ -3,7 +3,7 @@ package com.oskarro.justcode.repositories;
 import com.oskarro.justcode.domains.Article;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticleRepository extends CrudRepository<Article, Integer> {
+public interface ArticleRepository extends CrudRepository<Article, Long> {
 
 
 }
